requirements22.txt
aiodns==1.1.1
aiohttp==3.6.3
appdirs==1.4.4
argcomplete==1.12.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
ccxt==1.36.77
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
colorlog==4.6.2
constantly==15.1.0
cryptography==3.1.1
cryptowatch-sdk==0.0.14
dateparser==0.7.6
decorator==4.4.2
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
hyperlink==20.0.1
identify==1.5.11
idna==2.10
incremental==17.5.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.9.1
multidict==4.7.6
nodeenv==1.5.0
nox==2020.8.22
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.9.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.2
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pushbullet.py==0.11.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycoingecko==1.4.0
pycparser==2.20
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.1.0
wcwidth==0.2.5
websocket-client==0.57.0
yarl==1.1.0
zope.interface==5.1.2
BitstampClient
backtrader
